modify: board modify,delete

+++ 코어/enjoyCamp/src/main/java/com/ssafy/campcino/dto/requestDto/CreateReviewRequestDto.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class CreateReviewRequestDto {
     private Integer campId;
-    private Integer campRate;
+    private double campRate;
     private String comment;
     private String userId; // 사용자 ID
 }
